Refactor PMA context assembly
